backend/requirements.txt
flask==3.1.2
flask-cors==4.0.0
google-generativeai==0.8.5
python-dotenv==1.2.1
pandas==2.2.3
supabase==2.10.0
